Renamed the Modules from dnos* -> dellos*, updated copyright, removed dnos6_template
