+++ lib/ansible/module_utils/dellos6.py
@@ -1,3 +1,8 @@
+#
+# (c) 2015 Peter Sprygada, <psprygada@ansible.com>
+#
+# Copyright (c) 2016 Dell Inc.
+#
 # This code is part of Ansible, but is an independent component.
 # This particular file snippet, and this file snippet only, is BSD licensed.
 # Modules you write using this snippet, which is embedded dynamically by Ansible
@@ -27,47 +32,21 @@
 import re
 
 from ansible.module_utils.shell import CliBase
-from ansible.module_utils.network import register_transport, to_list, Command
-from ansible.module_utils.netcfg import NetworkConfig, ConfigLine
-
+from ansible.module_utils.network import Command, register_transport, to_list
+from ansible.module_utils.netcfg import NetworkConfig
 
 def get_config(module):
-    contents = module.params['config']
-
+    contents = module.params['running_config']
     if not contents:
-        contents = module.config.get_config()
-        module.params['config'] = contents
+        contents = module.cli('show running-config all')[0]
+        module.params['running_config'] = contents
 
-    return NetworkConfig(indent=1, contents=contents[0])
-
-
-def get_sublevel_config(running_config, module):
-    contents = list()
-    current_config_contents = list()
-    obj = running_config.get_object(module.params['parents'])
-    if obj:
-        contents = obj.children
-    contents[:0] = module.params['parents']
-
-    for c in contents:
-        if isinstance(c, str):
-            current_config_contents.append(c)
-        if isinstance(c, ConfigLine):
-            current_config_contents.append(c.raw)
-    sublevel_config = '\n'.join(current_config_contents)
-
-    return sublevel_config
+    return NetworkConfig(indent=1, contents=contents)
 
 
 class Cli(CliBase):
 
     NET_PASSWD_RE = re.compile(r"[\r\n]?password:\s?$", re.I)
-
-    WARNING_PROMPTS_RE = [
-        re.compile(r"[\r\n]?\[confirm yes/no\]:\s?$"),
-        re.compile(r"[\r\n]?\[y/n\]:\s?$"),
-        re.compile(r"[\r\n]?\[yes/no\]:\s?$")
-    ]
 
     CLI_PROMPTS_RE = [
         re.compile(r"[\r\n]?[\w+\-\.:\/\[\]]+(?:\([^\)]+\)){,3}(?:>|#) ?$"),
@@ -75,17 +54,18 @@ class Cli(CliBase):
     ]
 
     CLI_ERRORS_RE = [
-        re.compile(r"% ?Error: (?:(?!\bdoes not exist\b)(?!\balready exists\b)(?!\bHost not found\b)(?!\bnot active\b).)*$"),
+        re.compile(r"% ?Error"),
         re.compile(r"% ?Bad secret"),
         re.compile(r"invalid input", re.I),
         re.compile(r"(?:incomplete|ambiguous) command", re.I),
         re.compile(r"connection timed out", re.I),
+        re.compile(r"[^\r\n]+ not found", re.I),
         re.compile(r"'[^']' +returned error code: ?\d+"),
     ]
 
+
     def connect(self, params, **kwargs):
         super(Cli, self).connect(params, kickstart=False, **kwargs)
-        self.shell.send('terminal length 0')
 
 
     def authorize(self, params, **kwargs):
@@ -93,17 +73,13 @@ class Cli(CliBase):
         self.run_commands(
             Command('enable', prompt=self.NET_PASSWD_RE, response=passwd)
         )
+        self.run_commands('terminal length 0')
 
 
     def configure(self, commands, **kwargs):
         cmds = ['configure terminal']
-        cmdlist = list()
-        for c in to_list(commands):
-            cmd = Command(c, prompt=self.WARNING_PROMPTS_RE, response='yes')
-            cmdlist.append(cmd)
-        cmds.extend(cmdlist)
+        cmds.extend(to_list(commands))
         cmds.append('end')
-
         responses = self.execute(cmds)
         responses.pop(0)
         return responses
