+++ lib/ansible/plugins/action/dellos10_template.py
