+++ lib/ansible/plugins/action/dellos6_config.py
