+++ lib/ansible/plugins/action/dellos9_template.py
